567. Permuatation in String
Solves problem by utilizing two arrays that keep character counts of s1
and the sliding window of s2 and compares if the two arrays have 26
matches (all counts of every letter in the alphabet are equal). If other
characters besides lowercase english alphabet is used, a dictionary
(hashmap) would probably be better. This algorithm runs in O(n) time
complexity since it iterates through the array of s2 once.