20. Valid Parentheses
Solved by creating a dictionary (hash map) that maps each left side of a
bracket {,(,[ to the right side },),]. Loops over each character in s
and either adds it to the stack if it is a key, or pops the stack and
checks if the popped value is equal to the current character. If it is
not, returns false, otherwise returns true if stack is empty at end of
iterating through string.