739. Daily Temperatures
Solved by using a stack that holds previous temperatures. As it iterates
through the list of temperatures, stack will pop off any values that are
less than current and add the number of days since (current index minus
index of specific day) and add that to result. This runs in O(n) time
since it iterates through temperatures array once.