1046. Last Stone Weight
Create max heap (min heap with negative values). While heap is not
empty, push values back onto heap after subtracting the two if popped
values are not equal. After there is 1 or 0 elements in array, append 0
to end of array (so that becomes the answer if array is empty), and
return absolute value of stones[0] (if remaining value, need to make
positive again).