128. Longest Consecutive Sequence
This problem requires an algorithm that must run in O(n) time, so a
sorting algorithm on the array is not possible. Instead, create a set of
unique values of input, iterate through list checking there is no
currentval-1 in the set, and then increase by 1 checking if the next
consecutive value is in the set. When that fails, take max of current
longest consecutive value (returnVal) and newly calculated value minus
the start value. Once iterated through entire array, return max val.
This runs in O(n) time with O(n) space.