78. Subsets
Backtracking problem that wants to get all subsets of list passed in.
Will work in O(n^2) time since all subsets of a given list is
O(n^2). Created a dfs function that recurses with a decision to include
current value or not, in which case it pops value off of subset.